change file structure
