+++ page/src/components/index.js
@@ -1,15 +1,17 @@
-export { default as SideBar } from './SideBar';
-export { default as Header } from './Header';
-export { default as PageInfo } from './PageInfo';
-export { default as SaveSnackbar } from './SaveSnackbar';
-export { default as AssignmentBox } from './AssignmentBox';
-export { default as Problem } from './Problem';
-export { default as TimePicker } from './TimePicker';
-export { default as UseRequest } from './UseRequest';
-export { default as UserAnswer } from './UserAnswer';
-export { default as Loading } from './Loading';
-export { default as MarkdownEditor } from './MarkdownEditor';
-export { default as MarkdownViewer } from './MarkdownViewer';
-export { default as ScoreStats } from './ScoreStats';
-export { default as QuestionSelector } from './QuestionSelector';
-export { default as StudentSelector } from './StudentSelector';
+export { default as SideBar } from './main/SideBar';
+export { default as Header } from './main/Header';
+export { default as BottomPopup } from './shared/BottomPopup';
+export { default as Loading } from './shared/Loading';
+export { default as PageInfo } from './shared/PageInfo';
+export { default as TimePicker } from './shared/TimePicker';
+export { default as UseRequest } from './shared/UseRequest';
+export { default as AssignmentBox } from './sub/AssignmentBox';
+export { default as MarkdownEditor } from './sub/MarkdownEditor';
+export { default as MarkdownViewer } from './sub/MarkdownViewer';
+export { default as Problem } from './sub/Problem';
+export { default as QuestionSelector } from './sub/QuestionSelector';
+export { default as SaveSnackbar } from './sub/SaveSnackbar';
+export { default as ScoreStats } from './sub/ScoreStats';
+export { default as StudentPopup } from './sub/StudentPopup';
+export { default as StudentSelector } from './sub/StudentSelector';
+export { default as UserAnswer } from './sub/UserAnswer';
