+++ page/src/components/main/Header.js
@@ -1,18 +1,13 @@
 import React from 'react';
-import { makeStyles } from '@material-ui/core/styles';
-import { Grid, ButtonBase } from "@material-ui/core";
 import clsx from 'clsx';
-
-import ExitToAppIcon from '@material-ui/icons/ExitToApp';
-import IconButton from '@material-ui/core/IconButton';
-
-import MenuIcon from '@material-ui/icons/Menu';
-import AppBar from '@material-ui/core/AppBar';
-import Toolbar from '@material-ui/core/Toolbar';
-import Slide from '@material-ui/core/Slide';
-import useScrollTrigger from '@material-ui/core/useScrollTrigger';
 import axios from "axios";
 import { useHistory } from "react-router-dom";
+
+import { makeStyles } from '@material-ui/core/styles';
+import { Grid, ButtonBase, IconButton, AppBar, Toolbar, Slide, useScrollTrigger } from "@material-ui/core";
+import ExitToAppIcon from '@material-ui/icons/ExitToApp';
+import MenuIcon from '@material-ui/icons/Menu';
+
 
 const drawerWidth = 300;
