+++ page/src/components/shared/BottomPopup.js
@@ -1,23 +1,18 @@
 import React from 'react';
-import { Grid } from "@material-ui/core";
+import { useHistory } from "react-router-dom";
 
 import PropTypes from 'prop-types';
 import classNames from 'classnames';
-import Button from '@material-ui/core/Button';
+
+import { Grid, Button, IconButton, Snackbar, SnackbarContent } from "@material-ui/core";
+import { green, grey, amber } from '@material-ui/core/colors/green';
+import { withStyles } from '@material-ui/core/styles';
 import CheckCircleIcon from '@material-ui/icons/CheckCircle';
 import ErrorIcon from '@material-ui/icons/Error';
 import InfoIcon from '@material-ui/icons/Info';
 import CloseIcon from '@material-ui/icons/Close';
 import OpenInBrowser from '@material-ui/icons/OpenInBrowser'
-import green from '@material-ui/core/colors/green';
-import grey from '@material-ui/core/colors/grey';
-import amber from '@material-ui/core/colors/amber';
-import IconButton from '@material-ui/core/IconButton';
-import Snackbar from '@material-ui/core/Snackbar';
-import SnackbarContent from '@material-ui/core/SnackbarContent';
 import WarningIcon from '@material-ui/icons/Warning';
-import { withStyles } from '@material-ui/core/styles';
-import { useHistory } from "react-router-dom";
 
 const variantIcon = {
     success: CheckCircleIcon,
