+++ page/src/components/shared/Loading.js
@@ -1,6 +1,5 @@
 import React from 'react';
-import { Grid } from '@material-ui/core';
-import CircularProgress from '@material-ui/core/CircularProgress';
+import { Grid, CircularProgress } from '@material-ui/core';
 
 function Loading(props) {
     return (
