+++ page/src/components/sub/StudentSelector.js
@@ -1,20 +1,13 @@
 import React, { useEffect } from 'react';
 import PropTypes from 'prop-types';
 import clsx from 'clsx';
+import { Loading } from '../';
+
+import { Table, TableBody, TableCell, TableContainer, TableHead, TablePagination, TableRow, TableSortLabel, Toolbar, Typography, Paper, Checkbox } from '@material-ui/core';
 import { lighten, makeStyles } from '@material-ui/core/styles';
-import Table from '@material-ui/core/Table';
-import TableBody from '@material-ui/core/TableBody';
-import TableCell from '@material-ui/core/TableCell';
-import TableContainer from '@material-ui/core/TableContainer';
-import TableHead from '@material-ui/core/TableHead';
-import TablePagination from '@material-ui/core/TablePagination';
-import TableRow from '@material-ui/core/TableRow';
-import TableSortLabel from '@material-ui/core/TableSortLabel';
-import Toolbar from '@material-ui/core/Toolbar';
-import Typography from '@material-ui/core/Typography';
-import Paper from '@material-ui/core/Paper';
-import Checkbox from '@material-ui/core/Checkbox';
-import { Loading } from '.';
+import { green, red } from '@material-ui/core/colors';
+import CheckCircleOutlineIcon from '@material-ui/icons/CheckCircleOutline';
+import HighlightOffIcon from '@material-ui/icons/HighlightOff';
 
 function descendingComparator(a, b, orderBy) {
     if (b[orderBy] < a[orderBy]) {
@@ -44,14 +37,19 @@ function stableSort(array, comparator) {
 
 const headCells = [
     { 
-        id: 'questionNumber',
+        id: 'userNumber',
         disablePadding: true, 
-        label: '문제 번호' 
+        label: '학번' 
     },
     { 
-        id: 'questionId',
+        id: 'submitted',
         disablePadding: false, 
-        label: '문제 ID' 
+        label: '답안을 하나 이상 제출함' 
+    },
+    { 
+        id: 'marked',
+        disablePadding: false, 
+        label: '채점을 완료함' 
     }
 ];
 
@@ -69,7 +67,7 @@ function EnhancedTableHead(props) {
                         indeterminate={numSelected > 0 && numSelected < rowCount}
                         checked={rowCount > 0 && numSelected === rowCount}
                         onChange={onSelectAllClick}
-                        inputProps={{ 'aria-label': 'select all questions' }}
+                        inputProps={{ 'aria-label': 'select all students' }}
                     />
                 </TableCell>
                 {headCells.map((headCell) => (
@@ -139,11 +137,11 @@ const EnhancedTableToolbar = (props) => {
         >
             {numSelected > 0 ? (
                 <Typography className={classes.title} color="inherit" variant="subtitle1" component="div">
-                    {numSelected}개의 문제를 선택함
+                    {numSelected}명의 학생을 선택함
                 </Typography>
             ) : (
                     <Typography className={classes.title} variant="h6" id="tableTitle" component="div">
-                        채점할 문제 선택
+                        채점할 학생 선택
                     </Typography>
                 )}
         </Toolbar>
@@ -181,7 +179,7 @@ function QuestionSelector(props) {
     const [error, setError] = React.useState(false);
     const [emptyRows, setEmptyRows] = React.useState(0);
     const [order, setOrder] = React.useState('asc');
-    const [orderBy, setOrderBy] = React.useState('questionNumber');
+    const [orderBy, setOrderBy] = React.useState('userNumber');
     const [selected, setSelected] = React.useState([]);
     const [page, setPage] = React.useState(0);
     const [rowsPerPage, setRowsPerPage] = React.useState(10);
@@ -189,26 +187,42 @@ function QuestionSelector(props) {
     useEffect(() => {
         setRows(undefined);
         setError(false);
-        setEmptyRows(rowsPerPage - Math.min(rowsPerPage, props.assign.questions.length - page * rowsPerPage));
+        setEmptyRows(rowsPerPage - Math.min(rowsPerPage, props.assign.students.length - page * rowsPerPage));
         setOrder('asc');
-        setOrderBy('questionNumber');
+        setOrderBy('userNumber');
         setPage(0);
         setRowsPerPage(10);
-
+        
         try {
-            let questionNumber = 1;
             let rowData = [];
-            for (const question of props.assign.questions) {
-                rowData.push({
-                    questionNumber: questionNumber++,
-                    questionId: question.questionId
-                });
+            for (const userNumber of props.assign.students) {
+                if (props.selectedQues === undefined || props.selectedQues.length === 0) {
+                    rowData.push({
+                        userNumber: userNumber
+                    });
+                } else {
+                    const answers = props.answersDict[userNumber].answers;
+                    let submittedCount = 0;
+                    let markedCount = 0;
+                    for (const questionId of props.selectedQues) {
+                        const ansOfQues = answers.find(ans => ans.questionId === questionId);
+                        if (ansOfQues !== undefined) {
+                            ++submittedCount;
+                            markedCount += (ansOfQues.score === -1 ? 0 : 1);
+                        }
+                    }
+                    rowData.push({
+                        userNumber: userNumber,
+                        submitted: submittedCount > 0,
+                        marked: submittedCount === markedCount
+                    });
+                }
             }
             setRows(rowData);
         } catch (err) {
             setError(true);
         }
-    }, [props.assign]);
+    }, [props.assign, props.answersDict, props.selectedQues]);
 
     const handleRequestSort = (event, property) => {
         const isAsc = orderBy === property && order === 'asc';
@@ -218,7 +232,7 @@ function QuestionSelector(props) {
 
     const handleSelectAllClick = (event) => {
         if (event.target.checked) {
-            const newSelecteds = rows.map((n) => n.questionId);
+            const newSelecteds = rows.map((n) => n.userNumber);
             setSelected(newSelecteds);
             if (props.onChange)
                 props.onChange(newSelecteds);
@@ -229,12 +243,12 @@ function QuestionSelector(props) {
             props.onChange([]);
     };
 
-    const handleClick = (event, questionId) => {
-        const selectedIndex = selected.indexOf(questionId);
+    const handleClick = (event, userNumber) => {
+        const selectedIndex = selected.indexOf(userNumber);
         let newSelected = [];
 
         if (selectedIndex === -1) {
-            newSelected = newSelected.concat(selected, questionId);
+            newSelected = newSelected.concat(selected, userNumber);
         } else if (selectedIndex === 0) {
             newSelected = newSelected.concat(selected.slice(1));
         } else if (selectedIndex === selected.length - 1) {
@@ -260,15 +274,21 @@ function QuestionSelector(props) {
         setPage(0);
     };
 
-    const isSelected = (questionId) => selected.indexOf(questionId) !== -1;
+    const getIcon = (value) => {
+        if (value === undefined)
+            return "-";
+        else if (value)
+            return <CheckCircleOutlineIcon style={{ color: green[500] }} size="small"/>;
+
+        return <HighlightOffIcon style={{ color: red[500] }} size="small"/>;
+    }
+
+    const isSelected = (userNumber) => selected.indexOf(userNumber) !== -1;
 
     if (error)
-        return <Typography variant="h6" align="center">문제 정보를 가져올 수 없었습니다.</Typography>;
+        return <Typography variant="h6" align="center">답안 정보를 가져올 수 없었습니다.</Typography>;
     else if (rows === undefined)
-        return (
-            <Paper className={classes.paper} elevation={3}>
-                <Loading status="문제 정보를 가져오는 중..."></Loading>
-            </Paper>);
+        return <Typography variant="h6" align="center">채점할 문제들을 먼저 선택해주세요!</Typography>;
     else
         return (
             <Paper className={classes.paper} elevation={3}>
@@ -278,7 +298,7 @@ function QuestionSelector(props) {
                         className={classes.table}
                         aria-labelledby="tableTitle"
                         size="small"
-                        aria-label="question table"
+                        aria-label="students table"
                     >
                         <EnhancedTableHead
                             classes={classes}
@@ -293,17 +313,17 @@ function QuestionSelector(props) {
                             {stableSort(rows, getComparator(order, orderBy))
                                 .slice(page * rowsPerPage, page * rowsPerPage + rowsPerPage)
                                 .map((row, index) => {
-                                    const isItemSelected = isSelected(row.questionId);
-                                    const labelId = `question-table-checkbox-${index}`;
+                                    const isItemSelected = isSelected(row.userNumber);
+                                    const labelId = `student-table-checkbox-${index}`;
 
                                     return (
                                         <TableRow
                                             hover
-                                            onClick={(event) => handleClick(event, row.questionId)}
+                                            onClick={(event) => handleClick(event, row.userNumber)}
                                             role="checkbox"
                                             aria-checked={isItemSelected}
                                             tabIndex={-1}
-                                            key={row.questionId}
+                                            key={row.userNumber}
                                             selected={isItemSelected}
                                         >
                                             <TableCell padding="checkbox">
@@ -313,9 +333,14 @@ function QuestionSelector(props) {
                                                 />
                                             </TableCell>
                                             <TableCell component="th" id={labelId} scope="row" padding="none">
-                                                {row.questionNumber}
+                                                {row.userNumber}
                                             </TableCell>
-                                            <TableCell>{row.questionId}</TableCell>
+                                            <TableCell>
+                                                {getIcon(row.submitted)}
+                                            </TableCell>
+                                            <TableCell>
+                                                {getIcon(row.marked)}
+                                            </TableCell>
                                         </TableRow>
                                     );
                                 })}
