+++ page/src/shared/App.js
@@ -1,6 +1,8 @@
 import React, { Component } from 'react';
 import { Route, Switch, Redirect } from 'react-router-dom';
 import { Login, Main } from '../pages';
+import './pages.css';
+import './components.css';
 
 
 class App extends Component {
